+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+dev:
+  - clarify that --connection should be a URL
+
 1.15.0:
   - add --period to "synccommittee members", can be "current", "next"
   - add "validator expectation"